@@ -30,17 +30,13 @@ interface DBBuilderInterface
      * Process $func within a transaction. Any additional arguments after
      * $func are passed to $func as arguments
      *
-     * @param callable $func
-     * @param mixed ...$arguments
-     * @return mixed
-     *
      * @template TReturn
      * @psalm-param callable(mixed ...$arguments): TReturn $func
      * @psalm-return TReturn
      *
      * @throws DBException Common minimal exception thrown if anything goes wrong
      */
-    public function transaction(callable $func, ...$arguments);
+    public function transaction(callable $func, mixed ...$arguments): mixed;
 
     public function getDBInterface(): DBInterface;
 }
