@@ -9,7 +9,7 @@ trait DBPassToLowerLayerTrait
 {
     private DBRawInterface $lowerLayer;
 
-    public function transaction(callable $func, ...$arguments)
+    public function transaction(callable $func, mixed ...$arguments): mixed
     {
         return $this->lowerLayer->transaction($func, ...$arguments);
     }
@@ -19,7 +19,7 @@ trait DBPassToLowerLayerTrait
         return $this->lowerLayer->inTransaction();
     }
 
-    public function select($query, array $vars = []): DBSelectQueryInterface
+    public function select(string|array $query, array $vars = []): DBSelectQueryInterface
     {
         return $this->lowerLayer->select($query, $vars);
     }
@@ -34,46 +34,46 @@ trait DBPassToLowerLayerTrait
         $this->lowerLayer->clear($selectQuery);
     }
 
-    public function fetchOne($query, array $vars = []): ?array
+    public function fetchOne(string|array $query, array $vars = []): ?array
     {
         return $this->lowerLayer->fetchOne($query, $vars);
     }
 
-    public function fetchAll($query, array $vars = []): array
+    public function fetchAll(string|array $query, array $vars = []): array
     {
         return $this->lowerLayer->fetchAll($query, $vars);
     }
 
-    public function fetchAllAndFlatten($query, array $vars = []): array
+    public function fetchAllAndFlatten(string|array $query, array $vars = []): array
     {
         return $this->lowerLayer->fetchAllAndFlatten($query, $vars);
     }
 
     public function insert(
-        string $tableName,
+        string $table,
         array $row = [],
-        string $autoIncrementIndex = ''
+        string $autoIncrement = '',
     ): ?string {
-        return $this->lowerLayer->insert($tableName, $row, $autoIncrementIndex);
+        return $this->lowerLayer->insert($table, $row, $autoIncrement);
     }
 
     public function insertOrUpdate(
-        string $tableName,
+        string $table,
         array $row = [],
-        array $indexColumns = [],
-        ?array $rowUpdates = null
+        array $index = [],
+        ?array $update = null,
     ): void {
-        $this->lowerLayer->insertOrUpdate($tableName, $row, $indexColumns, $rowUpdates);
+        $this->lowerLayer->insertOrUpdate($table, $row, $index, $update);
     }
 
-    public function update(string $tableName, array $changes, array $where = []): int
+    public function update(string $table, array $changes, array $where = []): int
     {
-        return $this->lowerLayer->update($tableName, $changes, $where);
+        return $this->lowerLayer->update($table, $changes, $where);
     }
 
-    public function delete(string $tableName, array $where = []): int
+    public function delete(string $table, array $where = []): int
     {
-        return $this->lowerLayer->delete($tableName, $where);
+        return $this->lowerLayer->delete($table, $where);
     }
 
     public function change(string $query, array $vars = []): int
